refactor: create @bartools/types package and split report-service
- Create @bartools/types with shared enums, entity types, and API
  response types (33 tests). All packages now import from the single
  source of truth instead of maintaining separate copies.

- Split report-service.ts (576 LOC) into focused modules:
  - report-lifecycle.ts: create, photos, submit, progress
  - report-queries.ts: list, detail, stream state
  - report-review.ts: review workflow
  - report-record-helpers.ts: pure transform functions (12 tests)
  - catalog-queries.ts: bottle search, venue locations
  - report-service.ts: barrel re-export (zero consumer changes)

- Add satisfies checks in backend to enforce type contract at
  compile time against shared types.

- Migrate mobile to @bartools/types, delete packages/mobile/types/.

- Migrate dashboard schemas to use shared REPORT_STATUSES enum.
