+++ packages/backend/package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "@bartools/inference": "workspace:*",
+    "@bartools/types": "workspace:*",
     "drizzle-orm": "^0.45.2",
     "hono": "^4.12.12",
     "motia": "^1.0.4-rc.1",
