+++ packages/dashboard/package.json
@@ -11,6 +11,7 @@
     "test": "vitest run"
   },
   "dependencies": {
+    "@bartools/types": "workspace:*",
     "@emotion/react": "^11.14.0",
     "@mantine/core": "^9.0.1",
     "@mantine/hooks": "^9.0.1",
