--- packages/mobile/types/index.ts
@@ -1,18 +1,0 @@
-export type { ItemCategory, SessionStatus } from './enums';
-export {
-  ITEM_CATEGORIES,
-  SESSION_STATUSES,
-  FILL_LEVEL_MIN,
-  FILL_LEVEL_MAX,
-} from './enums';
-
-export type {
-  User,
-  Venue,
-  VenueMember,
-  Location,
-  Bottle,
-  Session,
-  Scan,
-  InventoryItem,
-} from './models';
