+++ packages/types/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@bartools/types",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "test": "bun test"
+  }
+}
